Add files via upload
The  BANKING SYSTEM  is an application for maintaining a person's account in a bank.  I tried to show the working of a banking account system and cover the basic functionality of a Bank Account Management System. To develop a project for solving financial applications of a customer in banking environment in order to nurture the needs of an end banking user by providing various ways to perform banking tasks. Also to enable the user’s work space to have additional functionalities which are not provided under a conventional banking project. 


The main aim of this project is to develop software to demonstrate the management of the banking system. This project is developed using Python’s tkinter module , python’s numpy module and database in python. This project  is designed as an interactive and content management system. The content management system deals with data entry, validation confirm and updating whiles the interactive system deals with system interaction with the administration and users.